test_dir/agent/requirements.txt
requests
PyYAML
psutil
py-cpuinfo
mss
Pillow
scapy
cryptography>=41.0.0
